fix variable type mismatch

+++ SlateTests/TestShellUtils.m
@@ -49,7 +49,7 @@
   NSString *result = [ShellUtils run:@"/bin/echo 'with single' \"and double quotes\"" wait:YES path:@"/"];
   NSError *err = nil;
   NSRegularExpression *testRegex = [NSRegularExpression regularExpressionWithPattern:@"with single and double quotes" options:0 error:&err];
-  int found = [testRegex numberOfMatchesInString:result options:0 range:NSMakeRange(0, [result length])];
+  unsigned long found = [testRegex numberOfMatchesInString:result options:0 range:NSMakeRange(0, [result length])];
   STAssertEquals(found, 1, @"Result should include all strings");
 }
